requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
paddlepaddle==3.0.0
paddleocr==2.9.1
python-multipart==0.0.17
pillow==11.0.0
